+++ src/AzureExtension/DataManager/IAzureDataManager.cs
@@ -42,6 +42,8 @@ public interface IAzureDataManager : IDisposable
 
     IEnumerable<Repository> GetDeveloperRepositories();
 
+    IEnumerable<PullRequests> GetPullRequestsForLoggedInDeveloperIds();
+
     // Repository name may not be unique across projects, and projects may not be unique across
     // organizations, so we need all three to identify the repository.
     PullRequests? GetPullRequests(string organization, string project, string repositoryName, string developerId, PullRequestView view);
