+++ src/AzureExtension/Widgets/Enums/WidgetDataState.cs
@@ -8,6 +8,7 @@ public enum WidgetDataState
     Unknown,
     Requested,      // Request is out, waiting on a response. Current data is stale.
     Okay,           // Received and updated data, stable state.
-    Failed,         // Failed retrieving data.
+    FailedUpdate,   // Failed updating data.
+    FailedRead,     // Failed to read the data.
     Disposed,       // DataManager has been disposed and should not be used.
 }
